+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/StartProcessResponse.java
@@ -26,7 +26,6 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.io.Serializable;
-import java.util.Map;
 import java.util.UUID;
 
 @JsonInclude(Include.NON_NULL)
@@ -34,22 +33,15 @@ public class StartProcessResponse implements Serializable {
 
     private final boolean ok = true;
     private final UUID instanceId;
-    private final Map<String, Object> out;
 
     @JsonCreator
-    public StartProcessResponse(@JsonProperty("instanceId") UUID instanceId,
-                                @JsonProperty("out") Map<String, Object> out) {
+    public StartProcessResponse(@JsonProperty("instanceId") UUID instanceId) {
 
         this.instanceId = instanceId;
-        this.out = out;
     }
 
     public UUID getInstanceId() {
         return instanceId;
-    }
-
-    public Map<String, Object> getOut() {
-        return out;
     }
 
     public boolean isOk() {
@@ -61,7 +53,6 @@ public class StartProcessResponse implements Serializable {
         return "StartProcessResponse{" +
                 "ok=" + ok +
                 ", instanceId=" + instanceId +
-                ", out=" + out +
                 '}';
     }
 }